todos, books, sync when offline

+++ app/src/main/kotlin/ru/boringowl/myroadmapapp/data/room/repos/UserRepository.kt
@@ -13,6 +13,7 @@ import retrofit2.HttpException
 import ru.boringowl.myroadmapapp.data.datastore.DataStorage
 import ru.boringowl.myroadmapapp.data.network.UserApi
 import ru.boringowl.myroadmapapp.data.network.errorText
+import ru.boringowl.myroadmapapp.data.room.AppDatabase
 import ru.boringowl.myroadmapapp.data.room.dao.UserDao
 import ru.boringowl.myroadmapapp.data.room.model.UserEntity
 import ru.boringowl.myroadmapapp.model.*
@@ -22,11 +23,11 @@ import javax.inject.Inject
 class UserRepository @Inject constructor(
     private val dao: UserDao,
     private val api: UserApi,
-    private val dataStorage: DataStorage
+    private val dataStorage: DataStorage,
+    private val db: AppDatabase,
 ) {
     private var isLoading by mutableStateOf(false)
 
-    private val dispUploader = DispUploader({ isLoading = true }, { isLoading = false })
 
     private fun entity(model: User) = UserEntity(model)
 
@@ -34,84 +35,71 @@ class UserRepository @Inject constructor(
         data: RegisterData,
         onSuccess: () -> Unit = {},
         onError: (message: String) -> Unit = {}
-    ) = dispUploader.load {
-        try {
-            val token = api.register(data)
-            dataStorage.setAuthToken(token.accessToken)
-            fetchMe()
-            onSuccess()
-        } catch (e: HttpException) {
-            onError(e.errorText())
-        }
+    ) = loadWithIO(
+        onNetworkError = onError,
+        onSuccess = onSuccess,
+    ) {
+        val token = api.register(data)
+        dataStorage.setAuthToken(token.accessToken)
+        fetchMe()
     }
 
     suspend fun login(
         data: LoginData,
         onSuccess: suspend () -> Unit = {},
         onError: suspend (message: String) -> Unit = {}
-    ) = dispUploader.load {
-        try {
-            val token = api.auth(data)
-            dataStorage.setAuthToken(token.accessToken)
-            fetchMe()
-            onSuccess()
-        } catch (e: HttpException) {
-            onError(e.errorText())
-        }
+    ) = loadWithIO(
+        onNetworkError = onError,
+        onSuccess = onSuccess,
+    ) {
+        val token = api.auth(data)
+        dataStorage.setAuthToken(token.accessToken)
+        fetchMe()
     }
 
     suspend fun resetPassword(
         data: RestorePasswordData,
         onSuccess: suspend () -> Unit = {},
         onError: suspend (message: String) -> Unit = {}
-    ) = dispUploader.load {
-        try {
-            api.resetPassword(data)
-            onSuccess()
-        } catch (e: HttpException) {
-            onError(e.errorText())
-        }
-    }
+    ) = loadWithIO(
+        onNetworkError = onError,
+        onSuccess = onSuccess,
+    ) { api.resetPassword(data) }
 
     suspend fun update(
         data: UserEmailData,
         onSuccess: suspend () -> Unit = {},
         onError: suspend (message: String) -> Unit = {}
-    ) = dispUploader.load {
-        try {
-            val user = api.updateEmail(data)
-            dao.insert(entity(user))
-            onSuccess()
-        } catch (e: HttpException) {
-            onError(e.errorText())
-        }
-    }
+    ) = loadWithIO(
+        onNetworkError = onError,
+        onSuccess = onSuccess,
+    ) { dao.insert(entity(api.updateEmail(data))) }
 
     suspend fun update(
         data: UpdatePasswordData,
         onSuccess: suspend () -> Unit = {},
         onError: suspend (message: String) -> Unit = {}
-    ) = dispUploader.load {
-        try {
-            api.updatePassword(data)
-            onSuccess()
-        } catch (e: HttpException) {
-            onError(e.errorText())
-        }
-    }
+    ) = loadWithIO(
+        onNetworkError = onError,
+        onSuccess = onSuccess,
+    ) { api.updatePassword(data) }
 
     suspend fun logout() {
         dao.delete()
         dataStorage.setAuthToken("")
+        db.flushDB()
     }
 
     fun get(): Flow<User?> = dao.get()
         .flowOn(Dispatchers.IO).conflate()
         .map { u -> u?.toModel() }
 
-    suspend fun fetchMe() = dispUploader.load {
+    suspend fun fetchMe() = loadWithIO {
         val user = api.me()
-        dao.delete()
-        dao.insert(entity(user))
+        val localUser = dao.getUser()?.toModel()
+        if (user != localUser) {
+            dao.delete()
+            dao.insert(entity(user))
+        }
     }
 }
